作成

+++ include/Tester/mpc/servo_motor_constraints_lti_mpc.hpp
@@ -1,18 +1,20 @@
 #ifndef __SERVO_MOTOR_CONSTRAINTS_LTI_MPC_HPP__
 #define __SERVO_MOTOR_CONSTRAINTS_LTI_MPC_HPP__
 
-#include "servo_motor_constraints_lti_mpc_lkf.hpp"
 #include "servo_motor_constraints_lti_mpc_F.hpp"
 #include "servo_motor_constraints_lti_mpc_Phi.hpp"
-#include "servo_motor_constraints_lti_mpc_solver_factor.hpp"
 #include "servo_motor_constraints_lti_mpc_Weight_U_Nc.hpp"
+#include "servo_motor_constraints_lti_mpc_lkf.hpp"
+#include "servo_motor_constraints_lti_mpc_solver_factor.hpp"
 
-#include "servo_motor_constraints_delta_U_min.hpp"
-#include "servo_motor_constraints_delta_U_max.hpp"
-#include "servo_motor_constraints_U_min.hpp"
+
 #include "servo_motor_constraints_U_max.hpp"
-#include "servo_motor_constraints_Y_min.hpp"
+#include "servo_motor_constraints_U_min.hpp"
 #include "servo_motor_constraints_Y_max.hpp"
+#include "servo_motor_constraints_Y_min.hpp"
+#include "servo_motor_constraints_delta_U_max.hpp"
+#include "servo_motor_constraints_delta_U_min.hpp"
+
 
 #include "python_mpc.hpp"
 
@@ -25,13 +27,17 @@ using namespace PythonMPC;
 constexpr std::size_t NP = 20;
 constexpr std::size_t NC = 2;
 
-constexpr std::size_t INPUT_SIZE = servo_motor_constraints_lti_mpc_lkf::INPUT_SIZE;
-constexpr std::size_t STATE_SIZE = servo_motor_constraints_lti_mpc_lkf::STATE_SIZE;
-constexpr std::size_t OUTPUT_SIZE = servo_motor_constraints_lti_mpc_lkf::OUTPUT_SIZE;
+constexpr std::size_t INPUT_SIZE =
+    servo_motor_constraints_lti_mpc_lkf::INPUT_SIZE;
+constexpr std::size_t STATE_SIZE =
+    servo_motor_constraints_lti_mpc_lkf::STATE_SIZE;
+constexpr std::size_t OUTPUT_SIZE =
+    servo_motor_constraints_lti_mpc_lkf::OUTPUT_SIZE;
 
 constexpr std::size_t AUGMENTED_STATE_SIZE = STATE_SIZE + OUTPUT_SIZE;
 
-constexpr std::size_t NUMBER_OF_DELAY = servo_motor_constraints_lti_mpc_lkf::NUMBER_OF_DELAY;
+constexpr std::size_t NUMBER_OF_DELAY =
+    servo_motor_constraints_lti_mpc_lkf::NUMBER_OF_DELAY;
 
 using LKF_Type = servo_motor_constraints_lti_mpc_lkf::type;
 
@@ -55,19 +61,19 @@ using Y_Min_Type = servo_motor_constraints_Y_min::type;
 
 using Y_Max_Type = servo_motor_constraints_Y_max::type;
 
-using PredictionMatrices_Type = MPC_PredictionMatrices_Type<
-  F_Type, Phi_Type, NP, NC, INPUT_SIZE, AUGMENTED_STATE_SIZE, OUTPUT_SIZE>;
+using PredictionMatrices_Type =
+    MPC_PredictionMatrices_Type<F_Type, Phi_Type, NP, NC, INPUT_SIZE,
+                                AUGMENTED_STATE_SIZE, OUTPUT_SIZE>;
 
-using Ref_Type = DenseMatrix_Type<double, OUTPUT_SIZE, 1>;
+using Ref_Type = DenseMatrix_Type<float, OUTPUT_SIZE, 1>;
 
-using ReferenceTrajectory_Type = MPC_ReferenceTrajectory_Type<
-  Ref_Type, NP>;
+using ReferenceTrajectory_Type = MPC_ReferenceTrajectory_Type<Ref_Type, NP>;
 
-using type = LTI_MPC_Type<
-  LKF_Type, PredictionMatrices_Type, ReferenceTrajectory_Type,
-  Weight_U_Nc_Type, Delta_U_Min_Type, Delta_U_Max_Type,
-  U_Min_Type, U_Max_Type, Y_Min_Type, Y_Max_Type,
-  SolverFactor_Type>;
+using type =
+    LTI_MPC_Type<LKF_Type, PredictionMatrices_Type, ReferenceTrajectory_Type,
+                 Weight_U_Nc_Type, Delta_U_Min_Type, Delta_U_Max_Type,
+                 U_Min_Type, U_Max_Type, Y_Min_Type, Y_Max_Type,
+                 SolverFactor_Type>;
 
 inline auto make() -> type {
 
@@ -83,19 +89,19 @@ inline auto make() -> type {
 
   auto delta_U_min = servo_motor_constraints_delta_U_min::make();
 
-  delta_U_min.template set<0, 0>(static_cast<double>(-100.0));
+  delta_U_min.template set<0, 0>(static_cast<float>(-100.0));
 
   auto delta_U_max = servo_motor_constraints_delta_U_max::make();
 
-  delta_U_max.template set<0, 0>(static_cast<double>(100.0));
+  delta_U_max.template set<0, 0>(static_cast<float>(100.0));
 
   auto U_min = servo_motor_constraints_U_min::make();
 
-  U_min.template set<0, 0>(static_cast<double>(-180.0));
+  U_min.template set<0, 0>(static_cast<float>(-180.0));
 
   auto U_max = servo_motor_constraints_U_max::make();
 
-  U_max.template set<0, 0>(static_cast<double>(180.0));
+  U_max.template set<0, 0>(static_cast<float>(180.0));
 
   auto Y_min = servo_motor_constraints_Y_min::make();
 
@@ -106,11 +112,10 @@ inline auto make() -> type {
   ReferenceTrajectory_Type reference_trajectory;
 
   auto lti_mpc = make_LTI_MPC(
-    kalman_filter, prediction_matrices, reference_trajectory, Weight_U_Nc,
-    delta_U_min, delta_U_max, U_min, U_max, Y_min, Y_max, solver_factor);
+      kalman_filter, prediction_matrices, reference_trajectory, Weight_U_Nc,
+      delta_U_min, delta_U_max, U_min, U_max, Y_min, Y_max, solver_factor);
 
   return lti_mpc;
-
 }
 
 } // namespace servo_motor_constraints_lti_mpc
